Hello World, we have a lot _todo_
This is _todo_ - a simple, lightweight todo list cli written in rust.
It is both, an exercise as well as a convenience cli replacing other
todo list apps like Apples Notes for example.

Its usage is inspired by the usage of git: `todo [cmd] [option] [args]`.

Run `$ todo help` to see the help menu and an overview of possible
commands.