refactor: handle errors
Until now we were not handling errors gracefully. For example, when an
argument is expected, e.g. when calling 'todo add _ ' we would not
handle a missing argument.

We hence pivot to using Option as argument input for the commands which
forces us to handle erros such as missing arguments (= None) gracefully.

Or at least more graceful than before.