feat(reword): allow reword to using the editor
If the flag --task, -m is not set for the 'reword' command, we now open
the user's default editor writing automatically the current task
description to it.

This allows the user to reuse the old description or write a completely
new one.

The --task, -m flag allows for a quick rewording.