refactor: use single database with several table | use config file
As of now, we have used one database with one table per (todo) list.

This commit changes this behavior and hence the backend logic. From now
on we will only have a single database and each todo list is implemented
as a table in that database. The tables are tracked by a "table of
tables" which consists of two columns: 'id' and 'name'.

The new approach is cleaner has the
following advantages:
- less verbose
- centralized backend
- offers the opportunity to easily query for all present lists.
- allows for easy generalization (for example setting up a database
  inside the cwd and a global one)

In order to know which is the currently checkedout table/todo list, we
introduce two variables in .env:
- CURRENT hodling the name of the current table
- PREVIOUS holding th ename of the previously checkedout table (we
  intedn to implement a "load -" method to checkout the previous table)

In addition we started using a dedicated config file (toml) which points to the
database. The config file will be stored using the XDG directories.

Remark: we decided to use toml as the language for the config file since
it is light-weight, intuitive to use and nicely integrates in the rust ecosystem.
After all cargo itslef uses toml.