test: add integration test for `init` command
We commence introducing integration tests starting with the `init`
command.

Integration tests are of course quite useful, especially if we (as we
certainly will) encounter future refactors.

In order to setup integration test, we need to refactor the `init`
command in such a way that we can "hijack" the location to the config
file, while in the test itself we provide a path to a temporary
directory and a mock config file.