fix(delete-list): don't allow to delete the list if currently in use
This commit is the start of fixing a bug that allowed us to delete
the list we were currently using. If removed succesfully we could still attempt to
write to the list (as it could be still checked out). This will
naturally fail with an error.

As it makes more sense to prevent the user to delete the list they are
currently using, we simply do so. Do not delete lists currently in use!

In addition, we synchronize our collection with our present tables by
deleting the entry (in our collection) corresponding to the list we just removed.