chore: add logs
To have more eyes on the code we add logs. We might have been going a
bit over the top with what to log but a) it's for training and b)
nothing better than a log log right?

A lot is for debugging as well (even though while adding the logs we
basically can debug by heart now - that is this version of the code).

Anyways, here goes nothing