fix: tag display and integration with sql
Previously, we used to err if no tag was provided. However, this causes the CLI to crash when trying to print the todo list, as the FromSql implementation for a missing tag would throw an error. We correct this behavior by simply writing an empty string if no tag was specified. Correspondingly, we adjust the Display implementation omittig the # when the tag is empty.