feat: list @due_date
We add the option to list all tasks with a given due date. The command
is `$ todo list @due_date` where `due_date` must be of the following
form
- today
- tomorrow
- weekday (3 letter abbriviation)
- dd-mm-yyyy

In order to make it work we needed to save 'today' as the midnight
timestamp. Howver, the 'created_at' coloumn still uses the exact timestamp.