refactor: improve structure of commands
Previously, all commands were implemented in the same file the TodoList
struct was declared. As we are adding more and more commands, this gets
more and more noisy and less maintainable.

In this commit we pivot to a more scalable structure of the code. Each
command has its own file in which they declare the implementation for
TodoList. These files are grouped by context.

For example, commands that performs the action of deletion, like
deleting a task, all tasks or a even an entire list, are grouped under
commands/delete.

As for naming commands, we decided for a semantic naming approach based on the
path. For example commands/delete/task deletes a task,
commands/delete/all_tasks deletes all tasts, etc.