refactor: refactor query logic
We refactor the query logic by deviding them into a Collection struct
and a Table struct which handle the user's collection (a table storing
the todo list id and its name) and the user's actual todo lists.

The struct play the role of the Rust layer above the SQL logic which
fetches the actual data from the SQL database and returns it.
Previously, the queries.rs file was building the SQL query and returning
the query string. The actual data retrieval was perfomred inside each
commands. The new architecture separates concerns and scales much better
than the previous approach.