refactor: implement repository pattern
We implement the repository pattern in order to
a) separate concerns cleanly: the commands of the CLI should not know
about which backend we are using. They should only implement domain
specific logic like add, list, etc.
b) make the code more maintainable: if we want to change or add another
backend in the future, we can now simply implement a new repository
c) make the code testable: we can mock a repository to easoly write unit
tests