tests(integration tests): write unit tests for sqlite repositories
We implement proper integration tests for the sqlite repositories

- todo_item_repository
- todo_list_repository

which handle the read and write inside a given todo list (i.e. a table
of todo tasks) and inside a collection (i.e. the table of todo lists).

The refatoring follows the following structure:

.
└── repository_integration
    ├── main.rs
    ├── mock
    │   ├── mock_sqlite.rs
    │   └── mod.rs
    └── sqlite
        ├── mod.rs
        ├── todo_item_repository.rs
        └── todo_list_repository.rs

'mock_sqlite.rs' implements utility structs and functions such as
mocking an in-memory db.