feat: hash id
Previously, we were using auto incremental integers as ids. While this
uniquely identifies tasks, it strikes us as not ideal. For example,
auto incremental integer ids are hard to maintain in import/export/sync
operations. Moreover they "leak" meta data such as
- the total number of tasks
- the temporal order the tasks were created
- database structure
We do not want to expose this kind of information to the user.

In this commit we hence pivot to hash id values, which do not suffer from these
issues. The id will be computed as
the hash of the fields
- task
- prio
- status
- due
- tag

This ties the actual content of the todo item closely to its id, similar
how git commit sha's are computed.

When listing the todo table, we display only the first 4 characters of
the id. We expect this to be enough to uniquely identify the task. Of
course, there is a non-zero probability that two todo items will have an
identical first 4 characters of their hashed id. In this case, we would
need to indicate more characters to uniquely identify the item. We will
come back to this problem in two ways:
1. we will implement a setting inside the configuration file which
   determines how many characters of the id we display
2. we will implement a 'show' command (similar to git) which displays
   the full id (and more detials such as meta data) of the corresponding
   todo item