refactor: move concerns of user paths to UserPaths struct
In this commit we move all responsibility related to paths to the
UserPaths struct. This gives us a single entry point to all operations
which depend on the paths to the database and config file.

In particular, we add more functionality to the UserPaths struct by
allowing it to expose the default paths.

We also make the CLI a bit less flexible in that we no longer allow the
user to change the paths of the configuration file. In any case, if the
user would like to change this, they could always create a symlink. The
benefit of this change is that the code is much more maintainable.