refactor: move init under src/commands/plumbing
We consider the init command as a plumbing command as it helps setting
up the CLI. In fact, the plumbing commands are no longer considered as
commands of the TodoList struct that have a self references, i.e. take
&self as an argument. They can hence be called without an instantiation
of the struct itself.

At this moment, we consider the following three commands as plumbing
commands (which are all used in the justfile)
- clean_data
- mod init
- mod show_paths