feat: add an update function
Previously, while we were able to update the status of an entry via
'close' and 'open', we were missing the option to update the due date, priority
or tag of an entry.

This commit implements this feature.

We update the previously defined 'update_status' function to a more
general 'update' function which takes the optional parameters 'due',
'prio', 'status' and 'tag'.

While working on this feature we used the chance to refactor the code
and improve the Datetime struct further. For example, we added a FromStr
functionality (we did the same for the Tag struct). We also moved the
util::parse_date to Datetime::parse. Our rational here is that the
Datetime struct should be responsible for all (due)date related logic,
such as parsing a date from a str.

Implementing the FromStr trait for Tag and Datetime also allows us to
use proper types (structs) when parsing commands and hence help improve
the code quality.

Finally we updated the integration tests to test the new backend query.