refactor(table styles): let table crate handle display of id
Previously, the list command was responsible for the display of the number of digits of the id. It did restrict the display of the id by chopping it up. However we find it more natural to transfer this responsibility to the table crate which should be responsible for all styling of cells.