refactor(Cmd enum): split plumbing and porcelain commands
We refactor the Cmd enum. We move it to a dedicated crate inside
`src/commands/domain` and add a new enum `Plumbing` which incorperates
all plumbing commands. All other commands are considered as "porcelain"
commands. For the latter, we need to initialize a todo list.
This split is conceptual and helps us distinguish the prerequisties
(like instantiate a TodoList, reading the .env file, etc) for the two
types of commands.