feat(commands/update): allow batch updates
Previously, we had to update tasks one by one.

This commit introduces the feature of batch updating. We have changed
the original 'update' command which now takes a list of ids as argument.
The change is applied to the 'close' and 'open' command accordingly.

Finally, we have updated the integration tests.