feat(tasks): support git-like split of title and message
We add support of splitting tasks into `title` and `message`. The title
is printed in the actual table, while `message` can be viewed using the
`show` command and the full task can be edited as before using the
`reword` command.

The advantage of this is that one can now split the task int subtask.
For example, a task like

"""
Title

[ ] task 1
[ ] task 2
"""

would now be shown as above via the `show` command while the table would
display onlye "Title" in the "task" column of the table.