feat(auto-completion): enable autocompletion
This commit enables auto-completion for the todo cli.

It separates the core logic and the actual cli into separate crates and
generates the auto-completions via a build.rs which runs at every build.

To install the auto-completions use `just install` (which installs the
release build entirely) or `just install-completions` to only install
the completions. This will run a scrip which detects the used shell and
moves the auto-completion file to the appropriate places. Depending on
the shell (bash, zsh or fish) the user might still have to update their
shell configs.