feat!(config): let todo.config handle the display of due and tag column
We add two boolean fields `shwo_due` and `show_tag` to the `todo.config` file
which handle the display of the corresponding columns, i.e. if set to
true, the column will be displayed in the `todo list` command; if false,
it will be hidden.

This allows the user to customize their UX.